apps/scraper/requirements.txt
kafka-python==2.0.2
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.2.2; sys_platform != 'win32'
python-dotenv==1.0.1